rework organization to isolate tests and libs
