+++ src/test/data/hexColors.js
