+++ src/test/data/rgbColors.js
