+++ src/test/ids.test.js
@@ -14,10 +14,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { groupTest } from './lib'
-import * as regex from './index'
-import { goodUsPhones, badUsPhones } from './test-data/usPhones'
-import { goodZipCodes, badZipCodes } from './test-data/zipCodes'
+/* globals test, expect */
+import * as regex from '../ids'
 
-groupTest(regex.usPhoneRE, goodUsPhones, badUsPhones)
-groupTest(regex.zipCodeRE, goodZipCodes, badZipCodes)
+const trivial0 = '00000000-0000-1000-8000-000000000000'
+const invalidUuid = '00000000-0000-0000-0000-000000000000'
+
+test('trivial (all 0) UUID is valid', () => {
+  expect(regex.uuidRE.test(trivial0)).toBe(true)
+})
+
+test('recognize invalid UUID', () => {
+  expect(regex.uuidRE.test(invalidUuid)).toBe(false)
+})
