+++ src/test/javascript.test.js
@@ -14,13 +14,19 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { groupTest } from './lib'
-import * as regex from './index'
-import { goodUrls, badUrls } from './test-data/urls'
-import { goodEmails, badEmails } from './test-data/emails'
+import * as regex from '../javascript'
+import { groupTest } from './lib/test-lib'
 
-groupTest(regex.urlRE, goodUrls, badUrls)
-groupTest(regex.emailRE, goodEmails, badEmails)
-groupTest(regex.emailEncodedOrNotRE,
-  goodEmails.map((e) => e.replace('@', '%40')),
-  badEmails.map((e) => e.replace('@', '%40')))
+groupTest(
+  regex.jsReservedWordRE,
+  ['for', 'in', 'var', 'yield'],
+  ['excited', 'javascript', 'ina'],
+  'JS reserved words'
+)
+
+groupTest(
+  regex.jsVariableRE,
+  ['blah', 'foo', 'ᾩ', 'ĦĔĽĻŎ', '〱〱〱〱', 'जावास्क्रिप्ट', 'KingGeorgeⅦ'],
+  ['for', 'in', 'var', 'yield', 'a-bad-name'],
+  'JS variable'
+)
