+++ src/test/lib/test-lib.js
@@ -16,11 +16,6 @@ limitations under the License.
 
 /* globals test, expect */
 
-export const lockdownRe = (str, flags) =>
-  Array.isArray(str)
-    ? new RegExp(`^\\s*(?:${str.join('|')})\\s*$`, flags)
-    : new RegExp(`^\\s*${str}\\s*$`, flags)
-
 const groupTestHelper = (re, data, isValid, desc) =>
   data.forEach((datum) =>
     test(`${desc ? desc + ' should ' : ''}${isValid ? 'pass' : 'fail'} '${datum}'`, () => {
